Step back (deleted Stack and Proxy) + Added ListAllocator (free list allocator)

+++ src/sgl/public/memory/pool_allocator.h
@@ -38,7 +38,7 @@ public:
 	/**
 	 * @copydoc Allocator::alloc
 	 */
-	virtual void * alloc(uint32 n);
+	virtual void * alloc(uint64 n);
 	
 	/**
 	 * @copydoc Allocator::free
